+++ pyflocks/util/command.py
@@ -1,57 +1,14 @@
-#!/usr/bin/python3
 import click
-import numpy as np
-
-from flock.factory import FlockFactory
-from util.geometry import EnumBounds, EnumNeighbours
-from util.plot     import *
-
-from typing import Any, Dict, List, Tuple
+from pyflocks.models.factory import FlockFactory
+from pyflocks.util.animate import __plot
 
 
-def __plot(
-        sim: 'FlockModel', t: int,
-        style: 'FlockStyle', traj: int, cmass: bool, sumvec: bool,
-        imgpath: str, simple: bool
-    ) -> None:
-
-    # bug when saving the first image, so save it again
-    if (t == 0):
-        prepare_state_plot(sim.l)
-        savefig(0, sim.title, sim.subtitle, imgpath)
-
-    coords = sim.traj['X'][t]
-    angles = sim.traj['A'][t]
-    # absolute velocity may not change in all models, so there is no trajectory
-    if 'V' in sim.traj.keys():
-        speeds = sim.traj['V'][t]
-    else:
-        speeds = sim.V
-
-    # plot phase and frequency of oscillator instead of angle and speed
-    if 'Kuramoto' in str(type(sim)) and style == FlockStyle.OSCIL:
-        angles = sim.traj['P'][t]
-        # absolute velocity may not change in all models, so there is no trajectory
-        if 'F' in sim.traj.keys():
-            speeds = sim.traj['F'][t]
-
-    if traj > 0:
-        plot_trajectories(t, sim.traj['X'], sim.l, traj, sim.bounds)
-
-    plot_state(style, t, sim.traj['X'][t], angles, speeds, sim.l, sim.dt, simple = simple)
-
-    if cmass:
-        if traj < 0:
-            traj = 20
-        plot_cmass(t, sim.traj['X'], sim.l, traj, sim.bounds)
-    if sumvec:
-        plot_sumvec(t, sim.traj['X'][t], angles, speeds, sim.l, sim.dt, simple = simple)
-
-    savefig(t, sim.title, sim.subtitle, imgpath, simple = simple)
+@click.Group()
+def cli():
+    pass
 
 
-
-@click.command()
+@cli.command()
 @click.option('--path', type = str, required = True,
               help = 'Directory with position data')
 @click.option('--out', type = str, required = True, default = 'out/img',
@@ -108,7 +65,3 @@ def plot(
         print(f"{t}: animating system to {imgpath}/")
         __plot(flock, t, FlockStyle[style], traj, cmass, sumvec, imgpath, simple)
         t += 1
-
-
-if __name__ == "__main__":
-    plot()
