+++ pyflocks/util/geometry.py
