+++ pyflocks/util/util.py
